third_party/WebKit/LayoutTests/html/marquee/marquee-shadow-root-no-access.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<marquee></marquee>
<script>
  var marquee = document.getElementsByTagName('marquee')[0];
  test(function() {
    assert_equals(marquee.shadowRoot, null);
  }, 'Marquee shadow-root should not be accessible');
</script>
